user register finished

+++ backend/models/weight_model.py
@@ -1,5 +1,5 @@
 from datetime import date
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Date
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Date, Float
 from sqlalchemy.orm import relationship
 from db import Base
 
@@ -8,7 +8,7 @@ class Weight(Base):
     __tablename__ = "weights"
 
     id = Column(String, primary_key=True, index=True)
-    weight = Column(Integer)
+    weight = Column(Float)
     magnitude = Column(String)
     date = Column(Date, default=date.today())
     user_id = Column(String, ForeignKey("users.id"))
